get ride of all the useless console.log, move 2 db function from polling/db to src/db
