+++ src/routes/dashboard.js
@@ -1,6 +1,6 @@
 const fetchCalls = require('../db/fetchCalls.js');
-const getFilterNameAndSpec = require('../../polling/db/getFilterNameAndSpec.js');
-const getTagNames = require('../../polling/db/getTagNamesArr.js');
+const getFilterNameAndSpec = require('../../src/db/getFilterNameAndSpec.js');
+const getTagNames = require('../../src/db/getTagNamesArr.js');
 const validate = require('../auth/validate.js');
 const pg = require('pg');
 const JWT = require('jsonwebtoken');
